Isolate RunningSummarizer into its own file
